pyproject.toml
[tool.poetry]
name = "cw7"
version = "0.1.0"
description = ""
authors = ["Lotos <shelkovnikova.da@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.3"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.10"
djangorestframework = "^3.15.2"
pillow = "^11.0.0"
django-filter = "^24.3"
djangorestframework-simplejwt = "^5.3.1"
ipython = "^8.30.0"
drf-yasg = "^1.21.8"
black = "^24.10.0"
setuptools = "^75.6.0"
stripe = "^11.3.0"
celery = "^5.4.0"
django-celery-beat = "^2.7.0"
redis = "^5.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
